3.2-crud/stocks_products/logistic/tests.py
from unittest import TestCase
from rest_framework.test import APIClient


class TestSomething(TestCase):
    def test_simple(self):
        client = APIClient
        resp = client.get('/api/v1/test/')
        self.assertEqual(resp.status_code, 200)
